+++ csrc/maths/idct.h
@@ -1,3 +1,4 @@
 #include <stdint.h>
+#include "complex.h"
 
-void idct2(const int16_t inp[64], uint8_t out[64]);
+void idct2(const int16_t inp[64], float out[64]);
